fix: refactor common function
